added modals and confetti for winner and loser
