+++ client/src/main.jsx
@@ -15,10 +15,10 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import App from './App';
 import Error from './pages/Error';
 import Game from './pages/GameBoard';
-import Comment from './pages/Comment';
+import Comment from './components/Comment';
 import Home from './pages/Home';
 import Login from './pages/Login';
-import Logout from './pages/Logout';
+// import Logout from './pages/Logout';
 import Auth from './utils/auth';
 import Instructions from './components/Instructions';
 
@@ -59,7 +59,7 @@ ReactDOM.render(
             <Route path="/game" element={<Game />} />
             <Route path="/comment" element={<Comment />} />
             {/* <Route path="/login" element={<Login />} /> */}
-            <Route path="/logout" element={<Logout />} />
+            {/* <Route path="/logout" element={<Logout />} /> */}
             <Route path="/instructions" element={<Instructions />} />
             <Route path="*" element={<Error />} />
           </Routes>
