+++ client/src/pages/GameBoard.jsx
@@ -1,10 +1,14 @@
 import InstructionsModal from "../components/InstructionsModal";
 import MyScoresModal from "../components/MyScoresModal";
-import LeaderBoardModal from "../components/LeaderBoardModal";
+import "../style.css";
+import Auth from '../utils/auth';
+
 // import Dropdown from "react-bootstrap/Dropdown";
 // import DropdownButton from "react-bootstrap/DropdownButton";
 // import Success from "../components/Success";
-// import Button from "react-bootstrap/Button";
+import Confetti from 'react-confetti';
+
+import Button from "react-bootstrap/Button";
 import "../assets/style/GameBoard.css";
 import { useState, useEffect } from "react";
 import "../assets/style/Timer.css";
@@ -12,10 +16,12 @@ import easyGames from "../puzzles/easy_sudoku.json";
 import medGames from "../puzzles/medium_sudoku.json";
 import hardGames from "../puzzles/hard_sudoku.json";
 import difficulty from "../pages/DifficultyLevel";
-// import { useLocation } from "react-router-dom";
-// import { Alert } from "react-bootstrap";
+import Modal from 'react-bootstrap/Modal';
+import "../assets/style/Modal.css";
+
 import React from "react";
 import { useStopwatch } from "react-timer-hook";
+import Comment from '../components/Comment';
 
 let solvedArray = [];
 let unsolvedArray = [];
@@ -38,6 +44,38 @@ const initial = [
 ];
 
 export default function Game() {
+  // const [lgshow, setlgShow] = useState(false);
+  // const handleClose = () => setlgShow(false);
+  // const handleShow = () => setlgShow(true);
+  const [lgShow, setWinnerShow] = useState(false);
+  const [smShow, setLoserShow] = useState(false);
+
+
+  const button2Style = {
+    margin: "0 10px", // Add margin between buttons
+    width: "180px",
+    border: 'none',
+    padding: '10px 20px',
+    textAlign: 'center',
+    textDecoration: 'none',
+    display: 'inline-block',
+    cursor: 'pointer',
+    borderRadius: '16px'
+  };
+  const [comment, setComment] = useState('');
+  const [updated, setUpdated] = useState(comment);
+
+  const handleChange = (event) => {
+    setComment(event.target.value);
+    // console.log(event.target.value);
+  }
+  const handleClick = () => {
+    setUpdated(comment);
+    console.log(comment);
+    let userComment = comment;
+
+  };
+
   const [sudokuArr, setSudokuArr] = useState(getDeepCopy(initial));
 
   function getDeepCopy(arr) {
@@ -147,15 +185,22 @@ export default function Game() {
     for (let x = 0; x < solvedArray.length; x++) {
       for (let y = 0; y < usergrid.length; y++) {
         if (solvedArray[x][y] !== usergrid[x][y]) {
-          alert(`YOU'RE BAD`);
+          setLoserShow(true)
           return false;
         }
       }
     }
-    alert(`Success! You scored: ${score}`);
+    setWinnerShow(true)
     return true;
   }
 
+  // function handlemodal({ seconds }, { minutes }, solvedArray, usergrid, {
+    // pause }){
+    // checkSudoku({ seconds }, { minutes }, solvedArray, usergrid, {
+      // pause,
+    // })
+    
+  // }
   function MyStopwatch() {
     const { seconds, minutes, isRunning, start, pause } = useStopwatch({
       autoStart: true,
@@ -167,7 +212,7 @@ export default function Game() {
             <span>{minutes}</span> minutes <span>{seconds}</span> seconds
           </h4>
         </div>
-        <button className="level">
+        <button className="btn btn-warning level">
           <label className="mx-1" for="difficulty">
             Difficulty Level:
           </label>
@@ -189,7 +234,7 @@ export default function Game() {
           </select>
         </button>
         <button
-          className="level"
+          className="btn btn-warning level"
           onClick={(e) =>
             checkSudoku({ seconds }, { minutes }, solvedArray, usergrid, {
               pause,
@@ -199,7 +244,7 @@ export default function Game() {
           Finished
         </button>
         <button
-          className="level"
+          className="btn btn-warning level"
           onClick={(e) => resetSudoku(randomIndex, solvedArray, unsolvedArray)}
         >
           Try Again
@@ -214,27 +259,15 @@ export default function Game() {
           <div className="game">
             <div className="position-relative">
               <div className="modalButtons position-absolute top-0 end-0">
+              </div>
+            </div>
+             <div className="position-absolute top-0 start-0 mx-3 mt-3"> 
                 <MyScoresModal />
                 <InstructionsModal />
-              </div>
-            </div>
-            <div className="position-absolute top-0 start-0">
-              <div className="signInContainer">
-                <button className="signInArea mt-2">
-                  <h6 className="signInToSave">Sign in to save scores!</h6>
-                  <a href="/Login">
-                    <button className="signIn">Sign In</button>
-                  </a>
-                  <a href="/Comment">
-                    <button className="signIn">Save</button>
-                  </a>
-                  <div>
-                  </div>
-                </button>
-              </div>
-            </div>
+              
+            </div> 
             <div className="game-header">
-              <h3 className="mt-2">Sudoku</h3>
+              <h3 className="mt-2 fs-1">Sudoku</h3>
               <table>
                 <tbody>
                   {[0, 1, 2, 3, 4, 5, 6, 7, 8].map((row, rIndex) => {
@@ -271,9 +304,124 @@ export default function Game() {
                 </tbody>
               </table>
               <div className="buttonContainer">
+        <>
+          
+      <Modal
+        size="lg"
+        show={lgShow}
+        onHide={() => setWinnerShow(false)}
+        aria-labelledby="example-modal-sizes-title-lg"
+      >
+        <Modal.Header closeButton>
+          <Modal.Title id="example-modal-sizes-title-lg" className="text-black fs-1">
+Congratulations! Your Score: {}      </Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          <div>
+            <div class="position-absolute top-50 start-50 translate-middle w-100">
+
+              {/* <div className='text-center mb-5'>
+
+<h1 className='fs-1 text-black '>Congratulations!!!</h1>
+<h3 className='text-black fs-3'>You finished the game! </h3>
+
+</div> */}
+              <div className='mx-2 input-group mb-3'>
+                <input className='form-control'
+                  type="text"
+                  placeholder="Type you comment here!"
+                  id="message"
+                  name="message"
+                  onChange={handleChange}
+                  value={comment}
+                />
+ <div class="input-group-append mx-2">
+    
+ {Auth.loggedIn() ? (
+    <button className="btn btn-warning mx-1" onClick={handleClick} type="button">Save</button>
+ ) : (  <a href="/Login"> <button
+  className="btn btn-warning mx-1"> 
+  Login
+ </button>
+</a>)}
+    <a href='/game'>
+      <button class="btn btn-warning mx-1" type="button">Play Again</button>    
+            </a>
+   <a href='/home'>
+    <button class="btn btn-warning mx-1" type="button">Home</button>
+    </a> 
+    
+
+  </div>
+              </div>
+            </div>
+            <div class="position-absolute bottom-0 start-00 translate-middle-x">
+<Confetti
+  width={window.innerWidth}
+  height={window.innerHeight}
+  numberOfPieces={600}
+/>
+             
+            </div>
+          </div>
+        </Modal.Body>      </Modal>
+    </>
+
+    <Modal
+        size="sm"
+        show={smShow}
+        onHide={() => setLoserShow(false)}
+        aria-labelledby="example-modal-sizes-title-lg"
+      >
+        <Modal.Header closeButton>
+          <Modal.Title id="example-modal-sizes-title-lg" className="text-black fs-1">
+Try Again!          </Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          <div>
+            <div class="position-absolute top-50 start-50 translate-middle w-100">
+
+
+              <div className='mx-2 input-group mb-3'>
+                <input className='form-control'
+                  type="text"
+                  placeholder="Type you comment here!"
+                  id="message"
+                  name="message"
+                  onChange={handleChange}
+                  value={comment}
+                />
+ <div class="input-group-append mx-2">
+    
+
+    <a href='/game'>
+      <button class="btn btn-warning mx-1" onClick={(e) => resetSudoku(randomIndex, solvedArray, unsolvedArray)} type="button">Try Again</button>    
+            </a>
+
+           
+   <a href='/home'>
+    <button class="btn btn-warning mx-1" type="button">Home</button>
+    </a> 
+    
+
+  </div>
+              </div>
+            </div>
+            <div class="position-absolute bottom-0 start-00 translate-middle-x">
+{/* <Confetti
+  width={window.innerWidth}
+  height={window.innerHeight}
+  numberOfPieces={600}
+/> */}
+             
+            </div>
+          </div>
+        </Modal.Body>      </Modal>
+    
                 <div>
                   <MyStopwatch />
                 </div>
+
               </div>
             </div>
           </div>
